게시글 등록(내용입력에 스마트에디터2),삭제

+++ arion/src/main/java/com/arion/app/group/board/service/BoardVO.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class BoardVO {
 
-	private long boardNo;			//게시글 번호
+	private int boardNo;			//게시글 번호
 	private String boardTitle;		//제목
 	private String boardContent;	//내용
 	private String employeeName;	//작성자 이름
